% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESSpackage.R
\name{AdaptMH}
\alias{AdaptMH}
\title{Adaptive Metropolis Hastings}
\usage{
AdaptMH(f, sigma, llk, n, N, cov_post)
}
\arguments{
\item{f}{initial state}

\item{sigma}{covariance matrix of f}

\item{llk}{log likelihood function}

\item{n}{number of iterations}

\item{N}{dimension of f}

\item{cov_post}{this is the theoretical covariance matrix of the posterior}
}
\value{
This function returns an n x N matrix where N is the dimension of f
}
\description{
This is a function named 'AdaptMH'
which implement the Metropolis Hastings algorithm proposed by Gareth and Jeffrey (2012). However, instead of
using proposals involving empirical estimate of covariance matrix, we input the known covariance of posterior distribution.
}
\details{
We need to load package "mvtnorm" before using this function
}
