% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{AdaptMH_full}
\alias{AdaptMH_full}
\title{Adaptive Metropolis Hastings using proposals involving empirical estimate of covariance}
\usage{
AdaptMH_full(f, sigma, llk, n, N, beta = 0.05)
}
\arguments{
\item{f}{initial state}

\item{sigma}{covariance matrix of f}

\item{llk}{log likelihood function}

\item{n}{number of iterations}

\item{N}{dimension of f}

\item{beta}{fixed constant used in proposal, the default value is 0.05}

\item{N}{dimension of f}
}
\value{
This function returns a list which contains an n x N matrix where N is the dimension of f and the corresponding log likelihood.
}
\description{
This is a function named 'AdaptMH_full'
which implement the Metropolis Hastings algorithm proposed by Gareth and Jeffrey (2012).
}
\details{
We need to load package "mvtnorm" before using this function
}
