% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cov_mat}
\alias{cov_mat}
\title{Covariance Matrix of Elliptical Slice Sampling}
\usage{
cov_mat(sig_var, l, x, N)
}
\arguments{
\item{sig_var}{signal variance}

\item{l}{lengthscale parameter}

\item{x}{feature dataset}

\item{N}{size of data x}
}
\value{
covariance matrix of f
}
\description{
This is an example function named 'cov_mat'
which gives the covariance matrix of f, the vector of latent variables.
}
