% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{NMH}
\alias{NMH}
\title{Neal's Metropolis Hastings (1999)}
\usage{
NMH(f, sigma, llk, n, stepsize)
}
\arguments{
\item{f}{initial state}

\item{sigma}{covariance matrix of f}

\item{llk}{log likelihood function}

\item{n}{number of iterations}

\item{stepsize}{this is the fixed number considered in proposal}
}
\value{
This function returns a list which contains an n x N matrix where N is the dimension of f and the corresponding log likelihood.
}
\description{
This is a function named 'NMH'
which implement the Metropolis Hastings algorithm adapted by Neal in 1999.
}
