% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ess}
\alias{ess}
\title{Elliptical Slice Sampling}
\usage{
ess(f, sigma, llk, n)
}
\arguments{
\item{f}{initial state}

\item{sigma}{covariance matrix of f}

\item{llk}{log likelihood function}

\item{n}{number of iterations}
}
\value{
an n x N matrix where N is the dimension of f
}
\description{
This is an example function named 'ess'
which implement Elliptical Slice Sampling algorithm.
}
